requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
lightgbm==4.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
shap==0.44.1
chart.js==4.4.0
plotly==5.17.0
Jinja2==3.1.2
joblib==1.3.2
mysql-connector-python==8.2.0
bcrypt==4.1.2
qrcode==7.4.2
Pillow==10.1.0
